mqtt_broker/requirements.txt
paho-mqtt>=1.6.1
numpy>=1.21.0
python-dotenv>=0.19.0 
pydantic>=2.0.0
